clean up n fixin's
Change stop order in pause.  Finally crushed the bug where old view
messages were still being delivered after the view change and thus
different messenger group.  Geebus.  But combined with a new view id in
genesis reconfigure section, this seems to have eliminated that entire
nightmare.  Smooth transitions now.
